Move couch structures inside a folder

+++ Sources/CouchDB/ConnectionProperties.swift
@@ -26,7 +26,7 @@ public struct ConnectionProperties {
     public let host: String
 
     /// Port number where CouchDB server is listening for incoming connections
-    public let port: Int16
+    public let port: Int
 
     /// Whether or not to use a secured connection
     public let secured: Bool
@@ -47,7 +47,7 @@ public struct ConnectionProperties {
     ///     - secured: Whether or not to use a secured connection.
     ///     - username: CouchDB admin username. Defaults to `nil`.
     ///     - password: CouchDB admin password. Defaults to `nil`.
-    public init(host: String, port: Int16, secured: Bool, username: String?=nil, password: String?=nil) {
+    public init(host: String, port: Int, secured: Bool, username: String?=nil, password: String?=nil) {
         self.host = host
         self.port = port
         self.secured = secured
